requirements.txt
boto==2.49.0
Django==2.1.3
django-filter==2.0.0
django-multiselectfield==0.1.8
django-storages==1.7.1
djangorestframework==3.9.0
Markdown==3.0.1
Pillow==5.3.0
psycopg2-binary==2.7.5
pytz==2018.7